# Allow any implementation of the MultivariatePolynomials API