--- a/src/SymbolicUtils.jl
+++ b/src/SymbolicUtils.jl
@@ -36,8 +36,9 @@ include("rule.jl")
 include("matchers.jl")
 
 # Convert to an efficient multi-variate polynomial representation
-import AbstractAlgebra.Generic: MPoly, PolynomialRing, ZZ, exponent_vector
-using AbstractAlgebra: ismonomial, symbols
+import MultivariatePolynomials
+const MP = MultivariatePolynomials
+import DynamicPolynomials
 export expand
 include("abstractalgebra.jl")