--- a/src/types.jl
+++ b/src/types.jl
@@ -867,7 +867,10 @@ istree(a::Pow) = true
 
 operation(a::Pow) = ^
 
-arguments(a::Pow) = [a.base, a.exp]
+# Use `Union` to avoid promoting the base and exponent to the same type.
+# For instance, if `a.base` is a multivariate polynomial and  `a.exp` is a number,
+# we don't want to promote `a.exp` to a multivariate polynomial.
+arguments(a::Pow) = Union{typeof(a.base), typeof(a.exp)}[a.base, a.exp]
 
 Base.hash(p::Pow, u::UInt) = hash(p.exp, hash(p.base, u))