The current implementation only supports the `MPoly` polynomial representation of AbstractAlgebra.
By using the MultivariatePolynomials, any implementation of multivariate polynomials can be used as long as it implements the API.
I choose to add a default with DynamicPolynomials which has a reasonable tradeoff between compile-time vs runtime.
On the other hand, [TypedPolynomials](http://github.com/JuliaAlgebra/TypedPolynomials.jl/) is usually faster but its compile time can become prohibitive if there are too many variables so it's best no to use it by default.
I added a benchmark.
Before this PR:
```
julia> goldstein_price();
  2.088 ms (19781 allocations: 855.57 KiB)
  6.695 ms (80770 allocations: 3.36 MiB)
  15.477 ms (202060 allocations: 8.39 MiB)
```
After this PR:
```
julia> goldstein_price();
  1.104 ms (7483 allocations: 394.23 KiB)
  2.431 ms (16981 allocations: 1019.19 KiB)
  6.018 ms (36690 allocations: 2.44 MiB)
```
The representation used by DynamicPolynomials is mostly similar to `AbstractAlgebra.MPoly` so I guess the different comes from the fact that `AbstractAlgebra` used BigInt while DynamicPolynomials automatically promotes depending on what you multiply it with (so `Int` in this case).
Another difference is that even if you create, a lot of variables, the monomials of DynamicPolynomials won't contain all of them.
There is not a ring ZZ like in AbstractAlgebra which forces all monomials to have many exponents. In DynamicPolynomials and TypedPolynomials, you automatically get more variables as you multiply and add things together but it starts very sparse.
I don't think we notice a difference in this benchmark but it might get noticeable as
https://github.com/JuliaSymbolics/SymbolicUtils.jl/blob/491ad806c8a21b75c11efd930fbce7af0de1fcfe/src/abstractalgebra.jl#L85-L86
will typically create a lot of variables with subexpressions only involving a small subset of them.


Edit: fixes #271