@shashi 2021-04-08
A few things I wanted to know more about:

- Going forward we're going to need GCD see #231 
- Does DynamicPolynomials.jl or one of the packages this would enable use dense array polynomial representation?
- Does this leave the whole system non-GPL? (MIT / BSD licenses are ok)
- Would it be a lot of work to also allow swapping in AbstractAlgebra.MPoly in place of these polynomials?