@blegat 2021-04-09
> Going forward we're going to need GCD see #231

I noticed you opened an issue some time ago about this (https://github.com/JuliaAlgebra/DynamicPolynomials.jl/issues/57) but we never replied, sorry about that. `gcd` is only implemented for univariate polynomials at the moment but we can add it for multivariate polynomials as well. By the way, this is also a blocking issue for simplifying our rational function (see https://github.com/JuliaAlgebra/MultivariatePolynomials.jl/blob/a58fa070cf4d82943702f0366b3c192cc2995cb5/src/rational.jl#L3).

> Does DynamicPolynomials.jl or one of the packages this would enable use dense array polynomial representation?

What do you mean by dense exactly ? Do you mean we store the coefficients for all the monomials up to some degree, even if they are zero (hence we don't need to store the exponents ?). This is implemented in https://github.com/JuliaDiff/TaylorSeries.jl with the `TaylorN` types which contains on `HomogeneousPolynomial` for each degree which contains just a vector of coefficients (and not the corresponding monomials since it's all of them). `TaylorN` does not implement the MultivariatePolynomials API but it wouldn't be hard to do if we focus on the part of the API needed for SymbolicUtils.

> Does this leave the whole system non-GPL? (MIT / BSD licenses are ok)

Yes, there are very few dependencies (MutableArithmetics, OrderedCollections and DataStructures) which are all licensed under the MIT license.

> Would it be a lot of work to also allow swapping in AbstractAlgebra.MPoly in place of these polynomials?

No, we just need to implement the MultivariatePolynomials API for `AbstractAlgebra.MPoly`. MultivariatePolynomials is really just an API for allowing software to use multivariate polynomials without having to rely on a particular representation/implementation. Using this API allows SymbolicUtils to support essentially any implementation at the cost of just a lightweight depencency. They you can switch the default at any time and the user can play around different implementation, finding the one that best suits its use case.